chatbot/requirements.txt
langchain
streamlit
openai
transformers
python-dotenv
duckduckgo-search
yfinance
tiktoken
plotly